+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Reorder the asset `preconnect` and remove unneeded `dns-prefetch` ([Issue 9](https://github.com/nhsx/nhsx-website/issues/9))
 - Terms and conditions content amendments
 - Update the default `BASE_URL` to include www.
+- Added simple Web App Manifest file
 
 ## 1.0.0 - 04/04/2019